Fix bug with splitting options
When supplying the @Option annotation, the bug used to require you to add an extra dash to every argument, apart from the first one (even arguments that didn't have a dash at all). This will fix that.
Note: this doesn't take into account escaped spaces in paths. That is for a future time.