Fix cyclic dependency
Elementary depends on Utilitary, but Utilitary's tests also depended on Elementary. This cyclic dependency caused issues when building and testing. Furthermore, this test dependency referred to an older version of Elementary, which caused junit dependency mismatch issues.

All this is now solved by introducing a new itest module for all tests in Utilitary that depended on Elementary.

Signed-off-by: CC007 <Coolcat_the_best2@Hotmail.com>